refactor(perf): move CPU/sched profiler into perf-self-profile
