+++ dial9-tokio-telemetry/src/telemetry/events.rs
@@ -1,82 +1,3 @@
-#[cfg(feature = "cpu-profiling")]
-use crate::telemetry::format::{CpuSampleEvent, WorkerId};
-#[cfg(feature = "cpu-profiling")]
-use dial9_core::buffer::{Encodable, ThreadLocalEncoder};
-use serde::Serialize;
-#[cfg(feature = "cpu-profiling")]
-use std::sync::Arc;
-
-/// What triggered a CPU sample.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
-pub enum CpuSampleSource {
-    /// Periodic CPU profiling sample (frequency-based).
-    CpuProfile = 0,
-    /// Context switch captured by per-thread sched event tracking.
-    SchedEvent = 1,
-}
-
-impl CpuSampleSource {
-    /// Decode from a raw `u8` wire value.
-    pub fn from_u8(v: u8) -> Self {
-        match v {
-            1 => Self::SchedEvent,
-            _ => Self::CpuProfile,
-        }
-    }
-}
-
-#[cfg(feature = "cpu-profiling")]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub(crate) struct ThreadName(Arc<str>);
-
-#[cfg(feature = "cpu-profiling")]
-impl ThreadName {
-    pub(crate) fn new(name: String) -> Self {
-        Self(name.into())
-    }
-
-    pub(crate) fn as_str(&self) -> &str {
-        self.0.as_ref()
-    }
-}
-
-/// Data for a CPU stack trace sample. Implements `Encodable` so samples can
-/// be written directly into the thread-local trace buffer without an
-/// intermediate enum. Interning of `thread_name` and `callchain` happens in
-/// the `Encodable::encode` impl.
-#[cfg(feature = "cpu-profiling")]
-#[derive(Debug, Clone)]
-pub(crate) struct CpuSampleData {
-    pub timestamp_nanos: u64,
-    pub worker_id: WorkerId,
-    pub tid: u32,
-    pub thread_name: Option<ThreadName>,
-    pub source: CpuSampleSource,
-    pub callchain: Vec<u64>,
-    /// CPU the sample was taken on, if the backend could determine it.
-    pub cpu: Option<u32>,
-}
-
-#[cfg(feature = "cpu-profiling")]
-impl Encodable for CpuSampleData {
-    fn encode(&self, enc: &mut ThreadLocalEncoder<'_>) {
-        let thread_name = self
-            .thread_name
-            .as_ref()
-            .map(|n| enc.intern_string(n.as_str()));
-        let callchain = enc.intern_stack_frames(&self.callchain);
-        enc.encode(&CpuSampleEvent {
-            timestamp_ns: self.timestamp_nanos,
-            worker_id: self.worker_id,
-            tid: self.tid,
-            source: self.source,
-            thread_name,
-            callchain,
-            cpu: self.cpu.map(u64::from),
-        });
-    }
-}
-
 /// Get the OS thread ID (tid) of the calling thread via `gettid()`.
 #[cfg(any(target_os = "linux", target_os = "android"))]
 pub(crate) fn current_tid() -> u32 {
@@ -253,51 +174,5 @@ mod tests {
             );
             prev = next;
         }
-    }
-}
-
-#[cfg(all(test, feature = "cpu-profiling"))]
-mod cpu_tests {
-    use crate::telemetry::analysis_events::Dial9Event;
-
-    /// Encode a single `CpuSampleData` through a real thread-local buffer
-    /// and decode it back via the `decode_events` path, asserting that
-    /// the `cpu` field round-trips.
-    fn cpu_sample_round_trip(cpu: Option<u32>) -> Dial9Event {
-        use crate::telemetry::buffer::encode_single;
-        use crate::telemetry::events::{CpuSampleData, CpuSampleSource};
-        use crate::telemetry::format::{WorkerId, decode_events};
-
-        let data = CpuSampleData {
-            timestamp_nanos: 12_345,
-            worker_id: WorkerId::from(0usize),
-            tid: 4242,
-            thread_name: None,
-            source: CpuSampleSource::CpuProfile,
-            callchain: vec![0xdead_beef, 0xcafe_babe],
-            cpu,
-        };
-        let encoded = encode_single(&data);
-        let events = decode_events(&encoded).expect("decode");
-        assert_eq!(events.len(), 1);
-        events.into_iter().next().unwrap()
-    }
-
-    #[test]
-    fn cpu_sample_event_round_trips_with_cpu() {
-        let Dial9Event::CpuSampleEvent(e) = cpu_sample_round_trip(Some(7)) else {
-            panic!("expected CpuSampleEvent");
-        };
-        assert_eq!(e.tid, 4242);
-        assert_eq!(e.cpu, Some(7));
-        assert_eq!(e.callchain, vec![0xdead_beef, 0xcafe_babe]);
-    }
-
-    #[test]
-    fn cpu_sample_event_round_trips_without_cpu() {
-        let Dial9Event::CpuSampleEvent(e) = cpu_sample_round_trip(None) else {
-            panic!("expected CpuSampleEvent");
-        };
-        assert_eq!(e.cpu, None);
     }
 }
